backend/requirements.txt
selenium
webdriver-manager
sqlalchemy
mysql-connector-python
python-dotenv
requests
fastapi
uvicorn
langchain
openai
beautifulsoup4
tqdm
faiss-cpu
python-dotenv
requests
pydantic
chromadb
lxml
langchain-community
sentence-transformers
faiss-cpu
transformers
accelerate
sentencepiece
bitsandbytes
pymysql
sqlalchemy
geopy
cryptography